+++ packages/celer-web-app/src/core/context/index.ts
@@ -1,5 +1,4 @@
 export * from "./AppExperimentsContext";
 export * from "./AppStateContext";
-export * from "./ServiceContext";
 export * from "./DocumentContext";
 export * from "./SettingContext";