+++ packages/celer-web-app/src/core/experiments.ts
@@ -16,7 +16,5 @@ export const useExpNoTrackDocPos = ():boolean => useAppExperiment("NoTrackDocPos
 export const useExpWarnNegativeVar = (): boolean => useAppExperiment("WarnNegativeVar", false);
 // Enable engine to infer coordinates for lines without coordinates (so that every line can be clicked)
 export const useExpInferCoord = ():boolean => useAppExperiment("InferCoord", true);
-// [08-11 Added] Use New DocumentContext and not use old ServiceContext. New DocumentContext will not work with leaflet map
-export const useExpNewDP = ():boolean => useAppExperiment("NewDP", true);
 // [08-25 Added] Use wasm bundler (only applies to dev server)
 export const useExpBetterBundler = ():boolean => useAppExperiment("BetterBundler", true);